remove data explorer
