+++ src/plugins/explore/public/application/legacy/discover/build_services.ts
@@ -52,12 +52,11 @@ import { UiActionsStart } from '../../../../../ui_actions/public';
 import { VisualizationsStart } from '../../../../../visualizations/public';
 import { SavedObjectOpenSearchDashboardsServices } from '../../../../../saved_objects/public';
 
-import { createSavedExploreLoader, SavedExplore } from '../../..';
+import { createSavedExploreLoader, DataExplorerServices, SavedExplore } from '../../..';
 import { getHistory } from './opensearch_dashboards_services';
 import { OpenSearchDashboardsLegacyStart } from '../../../../../opensearch_dashboards_legacy/public';
 import { UrlForwardingStart } from '../../../../../url_forwarding/public';
 import { NavigationPublicPluginStart } from '../../../../../navigation/public';
-import { DataExplorerServices } from '../data_explorer';
 import { Storage } from '../../../../../opensearch_dashboards_utils/public';
 import { DiscoverStartPlugins } from './types';
