+++ src/plugins/explore/public/components/sidebar/index.scss
