+++ src/plugins/explore/public/ui_metric/report_ui_metric.ts
@@ -4,7 +4,7 @@
  */
 
 import { UiStatsMetricType } from '@osd/analytics';
-import { METRIC_TYPE } from '../../../../../../usage_collection/public';
+import { METRIC_TYPE } from '../../../usage_collection/public';
 import { NEW_DISCOVER_APP_NAME } from './constants';
 import { getUsageCollector } from '../services';
