+++ src/plugins/explore/public/utils/state_management/hooks.ts
