+++ src/plugins/explore/public/utils/use/shallow_equal.ts
